[Update] Reorganize all the files for future updates
